streamlit-app/requirements.txt
boto3==1.34.80
botocore==1.34.103
docker==7.0.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
moto==5.0.7
numpy==1.24.4
packaging==23.2
pandas==2.1.4
pathlib==1.0.1
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
pydot==2.0.0
pylint==3.1.0
pytest==8.2.0
pytest-mock==3.14.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.1
scikit-image==0.22.0
scikit-learn==1.3.0
streamlit==1.32.2
urllib3==1.26.14
